+++ packages/stylelint/src/index.ts
@@ -2,8 +2,10 @@ import config from './standard'
 
 export default config
 export { config, config as standardConfig }
+export { MINIAPP_LANGUAGE_OPTIONS, STANDARD_LANGUAGE_OPTIONS } from './language-options'
 export { default as minPixelConfig } from './min-pixel'
 export { default as miniappConfig } from './miniapp'
 export { default as scssConfig } from './scss'
+export { mergeLanguageOptions } from './utils/merge-language-options'
 export { default as vueConfig } from './vue'
 export { default as vueScssConfig } from './vue-scss'