+++ packages/stylelint/src/scss.ts
@@ -1,6 +1,6 @@
 import type { Config } from 'stylelint'
 
 export default {
-  // https://github.com/stylelint-scss/stylelint-config-recommended-scss
-  extends: ['stylelint-config-recommended-scss'],
+  // https://github.com/stylelint-scss/stylelint-config-standard-scss
+  extends: ['stylelint-config-standard-scss'],
 } satisfies Config