+++ packages/stylelint/src/shared-rules/index.ts
@@ -1,0 +1,2 @@
+export * from './miniapp'
+export * from './vue'