+++ packages/stylelint/src/standard.ts
@@ -1,9 +1,12 @@
 import type { Config } from 'stylelint'
 import { IGNORE_FILES } from './constants'
+import { STANDARD_LANGUAGE_OPTIONS } from './language-options'
+import { mergeLanguageOptions } from './utils/merge-language-options'
 
 // https://stylelint.io/user-guide/options/
 export default {
   ignoreFiles: IGNORE_FILES,
+  languageOptions: mergeLanguageOptions(STANDARD_LANGUAGE_OPTIONS),
   extends: [
     // https://github.com/stylelint/stylelint-config-standard
     'stylelint-config-standard',
@@ -18,12 +21,21 @@ export default {
   // https://github.com/stylelint-stylistic/stylelint-config/blob/main/stylelint.config.js
   plugins: ['@stylistic/stylelint-plugin'],
   rules: {
-    'function-url-quotes': 'never',
-    'function-no-unknown': [true, { ignoreFunctions: ['constant'] }],
+    'function-url-quotes': 'always',
+    'function-no-unknown': [true, { ignoreFunctions: ['constant'] }], // 需与 languageOptions 配合使用
     // TODO: 对内联样式不起作用，如果要在两个属性之间自动添加空格，得用 Prettier 处理，临时解决方案是先 stylelint 再 prettier
     // https://github.com/stylelint-stylistic/stylelint-stylistic/blob/main/lib/rules/declaration-block-semicolon-space-after/README.md
     '@stylistic/declaration-block-semicolon-space-after': 'always-single-line',
     // https://github.com/stylelint-stylistic/stylelint-stylistic/blob/main/lib/rules/string-quotes/README.md#options
     '@stylistic/string-quotes': ['single', { avoidEscape: true }],
   },
+  overrides: [
+    {
+      files: ['**/*.html'],
+      rules: {
+        // HTML files default to unquoted url()
+        'function-url-quotes': 'never',
+      },
+    },
+  ],
 } satisfies Config