+++ packages/stylelint/tests/fixtures/uniapp/global.vue
@@ -1,0 +1,10 @@
+<template>
+  <view />
+</template>
+
+<style>
+page {
+  width: 20rpx;
+  height: 10px;
+}
+</style>